backend/requirements.txt
absl-py==0.11.0
asgiref==3.3.1
astunparse==1.6.3
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.3
branca==0.4.2
bs4==0.0.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
comtypes==1.1.8
cryptography==3.3.1
cycler==0.10.0
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.4
django-allauth==0.44.0
django-auth==0.1.9
django-cors-headers==3.6.0
django-import-export==2.5.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
et-xmlfile==1.0.1
flatbuffers==1.12
folium==0.12.1
futures==3.1.1
gast==0.4.0
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
goslate==1.5.1
grpcio==1.34.1
h5py==3.1.0
idna==2.10
img2pdf==0.4.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
jupyter-client==6.1.11
jupyter-core==4.7.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.3.4
MouseInfo==0.1.3
numpy==1.20.1
oauthlib==3.1.0
odfpy==1.4.1
opencv-python==4.5.1.48
openpyxl==3.0.7
opt-einsum==3.3.0
pandas==1.2.3
parso==0.8.1
phonenumbers==8.12.19
pickleshare==0.7.5
pikepdf==2.9.2
Pillow==8.0.1
prompt-toolkit==3.0.16
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///D:/virassistance/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
PyAutoGUI==0.9.52
pycodestyle==2.6.0
pycparser==2.20
PyDictionary==2.0.1
pygame==2.0.1
PyGetWindow==0.0.9
Pygments==2.8.0
PyJWT==1.7.1
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
python3-openid==3.2.0
pyttsx3==2.90
PyTweening==1.0.3
pytz==2020.5
pywhatkit==3.9
pywin32==300
PyYAML==5.4.1
pyzmq==22.0.3
reportlab==3.5.66
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.1
seaborn==0.11.1
six==1.15.0
soupsieve==2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
tablib==3.0.0
tb-nightly==2.5.0a20210308
tensorboard-data-server==0.3.0
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
tf-estimator-nightly==2.5.0.dev2021030801
tf-nightly==2.5.0.dev20210308
toml==0.10.2
torc==0.1.0
torch==1.8.0
torchaudio==0.8.0
torchvision==0.9.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
wikipedia==1.4.0
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.2.0
